requirements.txt
fastapi
uvicorn
crewai[tools]
langchain-openai
sse-starlette
python-dotenv
